refactor(oas31): standardize JSON Schema 2020-12 plugin extensions
Refs #8513
